requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.39.10
botocore==1.39.10
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
deprecation==2.1.0
easyocr==1.7.2
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
img2pdf==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lazy_loader==0.4
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ocrmypdf==16.10.4
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.0.0
pikepdf==9.10.2
pillow==11.3.0
pluggy==1.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.0.0
s3transfer==0.13.1
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
